+++ test/gauge_tests.cpp
@@ -8,6 +8,18 @@
 #include <cstdint>
 #include <ctime>
 
+#include <gauge/gauge.hpp>
+
+TEST(test_gauge, run_benchmarks)
+{
+    // Dummy arguments for gauge::runner
+    int argc = 1;
+    const char* argv[] = { "program" };
+
+    gauge::runner::add_default_printers();
+    gauge::runner::run_benchmarks(argc, argv);
+}
+
 GTEST_API_ int main(int argc, char** argv)
 {
     srand(static_cast<uint32_t>(time(0)));